Bug# 1042946 exec_relay_log_event memory leak in filered slave
There is a bug in deferred event container maintenance. When
the deferred events container rewind, the state "last_added" was
not reset and would point to some deleted event. However, the
memory address by the deleted event might be reused by new events
and would be thought to be fake "last_added". Therefore, they
would be missed for deletion and cause random memory leak over
time.

The fix is to reset "last_added" state to NULL when rewind
deferred events container.