Bug#16268289 LOCK_REC_VALIDATE_PAGE() MAY DEREFERENCE A POINTER TO A              FREED LOCK
ANALYIS
-------

In 5.5 code the lock_rec_block_validate() is called after releasing
the kernel mutex. There is a chance that the lock might be invalid so,
we are getting the valgrind error on invalid read on lock->index.

FIX
---

Fix would be to copy the lock->index when we are holding the kernel mutex 
and then pass it to the lock_rec_block_validate(). This implementation
is present in 5.1 code.  

[ Approved by sunny rb.no.oracle.com/rb/r/2152/ ]

--BZR--
revision-id: aditya.a@oracle.com-20130313061321-ibget09llt3x3ukn
property-audit-revid: aditya.a@oracle.com-20130313061321-v9ohvhd5bv0zhx54
property-branch-nick: mysql-5.5
testament3-sha1: c5e326f5771e32c026e6888f58e55b9e3325a1c5