Bug #16051728 SERVER CRASHES IN ADD_IDENTIFIER ON CONCURRENT ALTER TABLE AND SHOW ENGINE INNOD
Problem:

The purpose of explain_filename() is to provide useful additional
information regarding the partitions given the filename.  This function
was returning an error when it was not able to parse the given filename.
For example, within InnoDB, temporary files are created with #sql-
prefix.  But this function was not able to parse it correctly.

Solution:

It is not an error, if explain_filename() could not parse the given
filename.  If there is no partition information to explain, then silently
return from the function.

rb#1940 approved by mattiasj


--BZR--
revision-id: annamalai.gurusami@oracle.com-20130321061043-c5bosn82jwlcy8ee
property-audit-revid: annamalai.gurusami@oracle.com-20130321061043-opapkk24b6dy7y66
property-branch-nick: mysql-5.5
testament3-sha1: 400be392729685b878374a122e68b599bfaed1ed