Bug#16500013 : ADD VERSION CHECK TO MYSQL_UPGRADE
(Based on Sinisa's patch)

Added a version checking facility to mysql_upgrade.
The versions used for checking is the version of the
server that mysql_upgrade is going to upgrade and the
server version that mysql_upgrade was build/distributed
with.
Also added an option '--version-check' to enable/disable
the version checking.

--BZR--
revision-id: nirbhay.choubey@oracle.com-20130321172140-wlttfe3b2uv21w7g
property-audit-revid: nirbhay.choubey@oracle.com-20130321172140-hh0r8ocyuq2fl8o8
property-branch-nick: version_check-5.5
testament3-sha1: 37c1529107bda78e2817e86fd0de279d424f9930