Bug #13004581 	BLACKHOLE BINARY LOG WITH ROW IGNORES UPDATE AND DELETE STATEMENTS
When logging to the binary log in row, updates and deletes to a BLACKHOLE
engine table are skipped.
  
It is impossible to log binary log in row format for updates and deletes to
a BLACKHOLE engine table, as no row events can be generated in these cases.
After fix, generate a warning for UPDATE/DELETE statements that modify a
BLACKHOLE table, as row events are not logged in row format.
--BZR--
revision-id: bill.qu@oracle.com-20130427080454-xoewjixdlt09kqrs
property-audit-revid: bill.qu@oracle.com-20130427080454-wkpn5sh9kqfhpgvd
property-branch-nick: mysql-5.5
testament3-sha1: 06736fc084ec96ee26d2291587227a87b2694371