Bug #16411457  MASTER THREAD CANNOT EXIT FLUSH_LOOP WHEN INNODB_FAST_SHUTDOWN IS 2
Problem:

When innodb_fast_shutdown is set to 2 and the master thread enters
flush loop, under some circumstances it will not be able to exit it.
This may lead to a shutdown hanging.

This is happening because of the following:

1. In the flush_loop block of code, if the srv_fast_shutdown is
   equal to 2 (very fast shutdown), then we do not flush dirty
   pages in buffer pool to disk.
 
2. In the same flush_loop block of code, if the number of dirty
   pages is more than user specified limit, we go to step 1.

This results in infinite loop.

Solution:

When we are in the process of doing a very fast shutdown, don't
do step 2 above.

rb#2328 approved by Inaam.


--BZR--
revision-id: annamalai.gurusami@oracle.com-20130516043106-d76a3tgjck2dwq2n
property-audit-revid: annamalai.gurusami@oracle.com-20130516043106-nvwu9fyv0ga3k22c
property-branch-nick: mysql-5.5
testament3-sha1: 0115dd6e25def5e9e2bbcad6aa8da6149b79edc4