Bug #16417635 INNODB FAILS TO MERGE UNDER-FILLED PAGES DEPENDING ON DELETION ORDER
Problem:

When a InnoDB index page is under-filled, we will merge it with either
the left sibling node or the right sibling node.  But this checking is
incorrect.  When the left sibling node is available, even if merging
is not possible with left sibling node, we do not check for the 
possibility of merging with the right sibling node.  

Solution:

If left sibling node is available, and merging with left sibling node
is not possible, then check if merge with right sibling node is
possible.

rb#2506 approved by jimmy & ima.


--BZR--
revision-id: annamalai.gurusami@oracle.com-20130613054413-ytj3zynxf3ae34qn
property-audit-revid: annamalai.gurusami@oracle.com-20130613054413-gctzu0x6kbpfx3py
property-branch-nick: mysql-5.5
testament3-sha1: 2b14270905039951cd6c9a85365180d8cfea8d48