Bug#16914007-INNODB: CHECK TABLE SHOULD MARK AN INDEX AS CORRUPTED IF IT HAS A WRONG COUNT
If CHECK TABLE finds that a secondary index contains the wrong
number of entries, it used to report an error but not mark the
index as corrupt. The error means that the index should be rebuilt,
which can be done with ALTER TABLE DROP INDEX and ALTER TABLE ADD
INDEX.  But just in case the DBA does not pay any attention to the
output of CHECK TABLE, the secondary index should be marked as
corrupted so that it is not used again.

Approved by Inaam in RB:2607
--BZR--
revision-id: kevin.lewis@oracle.com-20130614183337-w17nidsrqp74lzdt
property-audit-revid: kevin.lewis@oracle.com-20130614183337-i66on9agcx93h1mx
property-branch-nick: mysql-5.5
testament3-sha1: 3be99a968d490fefe7872db27119c44d29472562