Bug#17033706 SINCE 5.5.32 & 5.6.12, INNODB CANT START WITH OWN              MULTI-FILE TABLESPACE
ANALYSIS
--------

When a tablespace has multiple data files, InnoDB fails to 
open the tablespace.  This is because for each ibd file, 
the first page is checked.But the first page of all ibd file
need not be the first page of the tablespace.  Only the first
page of the tablespace contains the tablespace header. When 
we check the first page of an ibd file that is not the first
page of the tablespace, then the "tablespace flags" is not
really available.This was wrongly used to check if a page is
corrupt or not.

FIX
---
Use the tablespace flags only if the page number is 0 
in a tablespace.  

[Approved by Inaam rb#2836 ]

--BZR--
revision-id: aditya.a@oracle.com-20130705090015-2ll580n3zbrmupwx
property-audit-revid: aditya.a@oracle.com-20130705090015-jhukvlh10uvkxuqk
property-branch-nick: mysql-5.5
testament3-sha1: 05aa717a9d7de16c692c03ac84c31ef6e25795bd