Bug #17168602 MYSQL_PLUGIN REMOVES NON-DIRECTORY TYPE FILES               SPECIFIED WITH THE BASEDIR OPTION
Description: The mysql_plugin client attempts to remove any
filename specified to the --basedir option. The problem is 
that if the filename does not end with a slash, it will 
attempt to unlink it, which succeeds for files, but not for 
directories.

Analysis: When we are starting mysql_plugin with basedir 
option and if we are giving path of a file as basedir, it 
deletes that file. It was because it uses a function 
my_delete which unlinks the file path given.

Fix:  As a fix we replace that line using another function 
my_free, which will only free the  pointer which is having 
that file path.

--BZR--
revision-id: arun.kuruvila@oracle.com-20130902092345-vhv2ol8x7rw7tbgv
property-audit-revid: arun.kuruvila@oracle.com-20130902092345-q3nn73l2yavwfvn0
property-branch-nick: fixed-5-5
testament3-sha1: 77320dd6b56d10e1df17935eb773ac3f12cafbe6