MYSQL-261: Innodb_mysql_master_log_pos appears to always be 1 event behind SHOW SLAVE STATUS
It is a bug in upstream code (bug#70564 opened). The problem seems to
be that when we commit we need to have the end binlog position. This is
done by setting future_group_master_log_pos in rli. But we set it at
wrong place i.e. in Query_log_event::do_apply_event.
At this point log_pos is not pointing to the end of current event.
We should instead update it
Xid_log_event::do_apply_event(Relay_log_info const *rli) which is
what we actually call when processing a commit.