WL#7266: Dump-thread additional concurrency tests
This worklog aims at testing the two following scenarios:

1) Whenever the mysql_binlog_send method (dump thread)
reaches the end of file when reading events from the binlog, before
checking if it should wait for more events, there was a test to
check if the file being read was still active, i.e, it was the last
known binlog. However, it was possible that something was written to
the binary log and then a rotation would happen, after EOF was
detected and before the check for active was performed. In this
case, the end of the binary log would not be read by the dump
thread, and this would cause the slave to lose updates.
This test verifies that the problem has been fixed. It waits during
this window while forcing a rotation in the binlog.

2) Verify dump thread can send events in active file, correctly after
encountering an IO error.
--BZR--
revision-id: nuno.carvalho@oracle.com-20131014144512-6yqeida5zw3qnn1e
property-audit-revid: nuno.carvalho@oracle.com-20131014144512-1uazqapos0y9ydyx
property-branch-nick: mysql-5.1
testament3-sha1: 1dd95f87bf0b8bd6fa6f617283188c92d818c68a