Bug #16051817  GOT ERROR 124 FROM STORAGE ENGINE                ON DELETE FROM A PARTITIONED TABLE
PROBLEM
-------

The user first disables all the non unique indexes
in the table and then rebuilds one partition.
During rebuild the indexes on that particular
partition are enabled. Now when we give a query 
the optimizer is unaware that on one partition 
indexes are enabled and if the optimizer selects
that index,myisam thinks that the index is not 
active and gives an error.

FIX
---

Before rebuilding a partition check whether non
unique indexes are disabled on the partitons.
If they are disabled then after rebuild disable
the index on the partition. 

[Approved by Mattiasj #rb3469]

--BZR--
revision-id: aditya.a@oracle.com-20131021063702-ak3uppt7h8a0czun
property-audit-revid: aditya.a@oracle.com-20131021063702-qcg06a9pxy9j1q9x
property-branch-nick: mysql-5.5
testament3-sha1: b0fb3981fbfd520b8719056b062f956b2c3fd33f