Bug #12917164 DROP USER CAN'T DROP USERS WITH LEGACY     UPPER CASE HOST NAME ANYMORE
Description:
It is not possible to drop users with host names with upper case
letters in them. i.e DROP USER 'root'@'Tmp_Host_Name'; is failing
with error.

Analysis: Since the fix 11748570 we came up with lower case hostnames
as standard. But in the current bug the hostname is created by
mysql_install_db script is still having upper case hostnames. 
So, if we have the hostname with upper case letters like(Tmp_Host_Name)
then we will have as it is stored in the mysql.user table. 
In this case if use "'DROP USER 'root'@'Tmp_Host_Name';" it gives 
error because we do compare with the lower case of hostname since the 
11748570 fix.

Fix: We need to convert the hostname to lower case before storing into 
the mysql.user table when we run the mysql_install_db script.
--BZR--
revision-id: venkata.sidagam@oracle.com-20131031173244-9vf8hy0y4jepgkcj
property-audit-revid: venkata.sidagam@oracle.com-20131031173244-h7thx1v5g03bc1bq
property-branch-nick: 5.1
testament3-sha1: 22188195049e2e1fbf41ee29c81c250e4b103a63